+++ Projects/Features/StorageFeature/Sources/Components/StorageComponent.swift
@@ -17,15 +17,7 @@ public protocol StorageDependency: Dependency {
 
 public final class StorageComponent: Component<StorageDependency>, StorageFactory {
     public func makeView() -> UIViewController {
-//        return StorageViewController.viewController(
-//            reactor: StorageReactor(),
-//            playlistStorageComponent: dependency.playlistStorageComponent,
-//            multiPurposePopUpFactory: dependency.multiPurposePopUpFactory,
-//            favoriteComponent: dependency.favoriteComponent,
-//            textPopUpFactory: dependency.textPopUpFactory,
-//            signInFactory: dependency.signInFactory
-//        )
-        return NewStorageViewController.viewController(
+        return StorageViewController.viewController(
             reactor: StorageReactor(),
             listStorageComponent: dependency.listStorageComponent,
             multiPurposePopUpFactory: dependency.multiPurposePopUpFactory,
